Jobs can declare keys to append
I think this works.  It probably needs more tests.

Two jobs that append the same keys will not overlap with eachother. But
if a job uses or updates a key, that will overlapp with an appender of
that key.

Appends are applied in job order.  So if the first job and second job
both append a key, the first job's appends will appear in the final
computed value before the second job's append.